+++ src/features/scoped-paths/toolkit/generate-excluded-paths-from-scoped-and-excluded-paths.spec.ts
@@ -1,5 +1,5 @@
 import assert from 'assert'
-import { DirectoryReader } from '../../../utils/types'
+import { DirectoryReader } from '../../../utils/directory-reader'
 import { generateExcludedPathsFromScopedAndExcludedPaths } from './generate-excluded-paths-from-scoped-and-excluded-paths'
 
 function createFakeDirectoryReader(paths: string[]): DirectoryReader {
