+++ src/features/scoped-paths/toolkit/get-target-paths-for-command.ts
@@ -1,5 +1,5 @@
 import { Uri, window } from 'vscode'
-import { uriToPath } from '../../../utils/uri'
+import { uriToPath } from '../../../utils/uri-to-path'
 
 /** Returns a list of the targeted paths for a VS Code command context */
 export function getTargetPathsForCommand(uri: Uri | undefined, selectedUris: Uri[] | undefined): string[] {
