+++ src/utils/directory-reader.ts
