--- src/utils/escape-regex.ts
@@ -1,3 +1,0 @@
-export function escapeRegex(input: string) {
-  return input.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
-}
