--- src/utils/is-multi-root-workspace.ts
@@ -1,7 +1,0 @@
-import * as vscode from 'vscode'
-
-export function isMultiRootWorkspace() {
-  return vscode.workspace.workspaceFolders
-    ? vscode.workspace.workspaceFolders.length > 1
-    : false
-}
