+++ src/utils/path-to-uri.ts
@@ -12,16 +12,3 @@ export function pathToUri(path: string): vscode.Uri | undefined {
 
   return vscode.Uri.joinPath(workspaceFolder.uri, path.slice(workspaceFolder.name.length + 1))
 }
-
-export function uriToPath(uri: vscode.Uri | undefined): string | undefined {
-  if (!uri) return undefined
-
-  const exactWorkspaceFolder = vscode.workspace.workspaceFolders?.find(workspaceFolder => workspaceFolder.uri.path === uri.path)
-  if (exactWorkspaceFolder) {
-    return exactWorkspaceFolder.name
-  }
-
-  const path = vscode.workspace.asRelativePath(uri, true)
-  // If workspace folder not found, it returns an absolute path
-  return path.startsWith('/') ? undefined : path
-}
